+++ src/main/java/teamkiim/koffeechat/domain/admin/corp/controller/CorpAdminApiDocument.java
@@ -1,4 +1,4 @@
-package teamkiim.koffeechat.domain.corp.controller;
+package teamkiim.koffeechat.domain.admin.corp.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -6,7 +6,6 @@ import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
