+++ src/main/java/teamkiim/koffeechat/domain/admin/corp/controller/request/CreateApprovedCorpRequest.java
@@ -1,4 +1,4 @@
-package teamkiim.koffeechat.domain.corp.controller.dto.request;
+package teamkiim.koffeechat.domain.admin.corp.controller.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
